Obtain grid fieldNames from screen metadata instead of from grid data passed into pui.show()
If a grid was loaded for the first time via replaceRecords(), it didn't work the way the code was
previously written, because when pui.show() was initially called with no grid data, the fieldNames
array was created empty. Then when replaceRecords() was called with data, a blank grid was
displayed, due to the data fieldnames not matching with the Grid's list of fieldnames (which was
empty).