This change is from Krista Pedersen <kpedersen@profoundlogic.com> - I am committing it as a favor to her.
Added a new "show week number" property to the date field widget, when
this is enabled, it will show the ISO week number to the left of the
days in the date field's popup calendar.

Also added a pui["week number calculator"] config option to allow
customers to define their own logic to calculate the week number in case
the ISO week number is not desired.